+++ packages-legacy/helux-preact/src/index.ts
@@ -18,6 +18,7 @@ export const {
   runDeriveTask,
   // watch api
   watch,
+  watchEffect,
   // hooks api
   useAtom,
   useAtomX,
@@ -25,6 +26,7 @@ export const {
   useReactiveX,
   useDerived,
   useWatch,
+  useWatchEffect,
   useGlobalId,
   useService,
   useOnEvent,