+++ packages/helux-hooks/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @helux/hooks
 
+## 4.0.3
+
+### Patch Changes
+
+- @helux/hooks-impl@4.0.3
+- @helux/types@4.0.3
+- @helux/utils@4.0.3
+
 ## 4.0.2
 
 ### Patch Changes